+++ Telerik.Sitefinity.Frontend.Forms/InterfaceMappings.cs
@@ -6,8 +6,8 @@ using Telerik.Sitefinity.Frontend.Forms.Mvc.Models.Fields.DropdownListField;
 using Telerik.Sitefinity.Frontend.Forms.Mvc.Models.Fields.FileField;
 using Telerik.Sitefinity.Frontend.Forms.Mvc.Models.Fields.MultipleChoiceField;
 using Telerik.Sitefinity.Frontend.Forms.Mvc.Models.Fields.ParagraphTextField;
-using Telerik.Sitefinity.Frontend.Forms.Mvc.Models.Fields.RecaptchaField;
-using Telerik.Sitefinity.Frontend.Forms.Mvc.Models.Fields.SectionHeaderField;
+using Telerik.Sitefinity.Frontend.Forms.Mvc.Models.Fields.Recaptcha;
+using Telerik.Sitefinity.Frontend.Forms.Mvc.Models.Fields.SectionHeader;
 using Telerik.Sitefinity.Frontend.Forms.Mvc.Models.Fields.SubmitButton;
 using Telerik.Sitefinity.Frontend.Forms.Mvc.Models.Fields.TextField;
