+++ Telerik.Sitefinity.Frontend.Forms/Mvc/Models/Fields/SectionHeader/SectionHeaderFieldViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace Telerik.Sitefinity.Frontend.Forms.Mvc.Models.Fields.SectionHeaderField
+﻿namespace Telerik.Sitefinity.Frontend.Forms.Mvc.Models.Fields.SectionHeader
 {
     /// <summary>
     /// This class represents view model for section header field.
