+++ Telerik.Sitefinity.Frontend.Forms/Mvc/Views/Recaptcha/DesignerView.Simple.cshtml
