+++ Telerik.Sitefinity.Frontend.Forms/Mvc/Views/SectionHeader/DesignerView.Simple.json
